Revert "refactor: modify plugin class loading order to follow parent delegation mechanism (#7258)"
This reverts commit 1d8a25cd69ae5241ff942ba30273049b1dcd84d5.
