+++ application/src/main/java/run/halo/app/plugin/DevPluginLoader.java
@@ -1,12 +1,10 @@
-package run.halo.app.plugin.loader;
+package run.halo.app.plugin;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
 import org.pf4j.DevelopmentPluginLoader;
-import org.pf4j.PluginClassLoader;
 import org.pf4j.PluginDescriptor;
 import org.pf4j.PluginManager;
-import run.halo.app.plugin.PluginProperties;
 
 public class DevPluginLoader extends DevelopmentPluginLoader {
 
@@ -41,12 +39,5 @@ public class DevPluginLoader extends DevelopmentPluginLoader {
     public boolean isApplicable(Path pluginPath) {
         // Currently we only support a plugin loading from directory in dev mode.
         return Files.isDirectory(pluginPath);
-    }
-
-    @Override
-    protected PluginClassLoader createPluginClassLoader(Path pluginPath,
-        PluginDescriptor pluginDescriptor) {
-        return new HaloPluginClassLoader(this.pluginManager, pluginDescriptor,
-            this.getClass().getClassLoader());
     }
 }
